feat(#313): remove CrewAI — Agent SDK becomes the production runtime
ADR-0006 Phase 2 complete. Closes epic #308.

What changed:
  - src/agent_sdk/_shared.py: extracts the framework-agnostic helpers
    (build_research_brief, audit_article_stats, apply_editorial_fixes,
    parse_research_for_verification, GRAPHICS_AGENT_PROMPT) that used
    to live in src/crews/stage3_crew.py and stage4_crew.py
  - src/agent_sdk/stage3_runner.py + stage4_runner.py: now import from
    _shared, no longer depend on src/crews
  - src/economist_agents/flow.py: rewritten as sequential Python on top
    of src.agent_sdk.pipeline.run_pipeline. The CrewAI Flow @start /
    @listen / @router decorators are gone.
  - src/crews/: deleted (stage3_crew, stage4_crew, development_crew,
    sprint_orchestrator_crew). The latter two had never run successfully
    in production per the 2026-04-21 sprint memo.
  - src/tools/research_tools.py + tests/test_research_tools.py: deleted.
    Agent SDK uses the web-researcher MCP server instead.
  - requirements.txt: crewai and crewai-tools removed.
  - tests: 91 tests passing on the migrated path (test_stat_audit +
    test_stage4_editorial_fixes + test_pre_commit_arch_check).
  - ADR-0006 status moved to Accepted with full timeline.
