+++ logs/article_evals.json
@@ -19,5 +19,89 @@
       "structure": "1 headings, 847 words, references: yes",
       "visual_engagement": "image: yes, chart embedded: no"
     }
+  },
+  {
+    "article_filename": "published",
+    "timestamp": "2026-04-16T23:37:36.170968",
+    "scores": {
+      "opening_quality": 4,
+      "evidence_sourcing": 3,
+      "voice_consistency": 10,
+      "structure": 1,
+      "visual_engagement": 5
+    },
+    "total_score": 23,
+    "max_score": 50,
+    "percentage": 46,
+    "details": {
+      "opening_quality": "Opening has 0 data tokens",
+      "evidence_sourcing": "0 references cited, 0 placeholders; fresh citations (2026/2025): 0; analyst vendors cited: 0",
+      "voice_consistency": "Clean voice",
+      "structure": "0 headings, 3 words, missing: ['layout', 'title', 'date', 'categories', 'image'], references: no",
+      "visual_engagement": "image: no, chart embedded: no"
+    }
+  },
+  {
+    "article_filename": "published",
+    "timestamp": "2026-04-17T00:11:55.680005",
+    "scores": {
+      "opening_quality": 4,
+      "evidence_sourcing": 3,
+      "voice_consistency": 10,
+      "structure": 1,
+      "visual_engagement": 5
+    },
+    "total_score": 23,
+    "max_score": 50,
+    "percentage": 46,
+    "details": {
+      "opening_quality": "Opening has 0 data tokens",
+      "evidence_sourcing": "0 references cited, 0 placeholders; fresh citations (2026/2025): 0; analyst vendors cited: 0",
+      "voice_consistency": "Clean voice",
+      "structure": "0 headings, 3 words, missing: ['layout', 'title', 'date', 'categories', 'image'], references: no",
+      "visual_engagement": "image: no, chart embedded: no"
+    }
+  },
+  {
+    "article_filename": "published",
+    "timestamp": "2026-04-17T10:02:56.858165",
+    "scores": {
+      "opening_quality": 4,
+      "evidence_sourcing": 3,
+      "voice_consistency": 10,
+      "structure": 1,
+      "visual_engagement": 5
+    },
+    "total_score": 23,
+    "max_score": 50,
+    "percentage": 46,
+    "details": {
+      "opening_quality": "Opening has 0 data tokens",
+      "evidence_sourcing": "0 references cited, 0 placeholders; fresh citations (2026/2025): 0; analyst vendors cited: 0",
+      "voice_consistency": "Clean voice",
+      "structure": "0 headings, 3 words, missing: ['layout', 'title', 'date', 'categories', 'image'], references: no",
+      "visual_engagement": "image: no, chart embedded: no"
+    }
+  },
+  {
+    "article_filename": "published",
+    "timestamp": "2026-04-17T17:10:29.400953",
+    "scores": {
+      "opening_quality": 4,
+      "evidence_sourcing": 3,
+      "voice_consistency": 10,
+      "structure": 1,
+      "visual_engagement": 5
+    },
+    "total_score": 23,
+    "max_score": 50,
+    "percentage": 46,
+    "details": {
+      "opening_quality": "Opening has 0 data tokens",
+      "evidence_sourcing": "0 references cited, 0 placeholders; fresh citations (2026/2025): 0; analyst vendors cited: 0",
+      "voice_consistency": "Clean voice",
+      "structure": "0 headings, 3 words, missing: ['layout', 'title', 'date', 'categories', 'image'], references: no",
+      "visual_engagement": "image: no, chart embedded: no"
+    }
   }
 ]
