--- src/crews/sprint_orchestrator_crew.py
@@ -1,855 +1,0 @@
-#!/usr/bin/env python3
-"""
-Sprint Orchestrator Crew - Manages entire sprint execution using development crews
-
-Coordinates multiple development crews working in parallel on different stories,
-following the established sprint management patterns.
-
-Architecture:
-    1. Sprint Planning - Parse SPRINT.md and validate Definition of Ready
-    2. Story Routing - Route each story to appropriate specialist crews
-    3. Parallel Execution - Coordinate multiple development crews
-    4. Quality Integration - Ensure all stories meet quality gates
-    5. Sprint Completion - Generate metrics and close sprint
-
-Usage:
-    from src.crews.sprint_orchestrator_crew import SprintOrchestratorCrew
-
-    crew = SprintOrchestratorCrew()
-    result = crew.execute_sprint(sprint_number=15)
-"""
-
-from concurrent.futures import ThreadPoolExecutor, as_completed
-from pathlib import Path
-from typing import Any
-
-from crewai import Agent, Crew, Task
-from crewai.process import Process
-
-from scripts.agent_registry import AgentRegistry
-from src.crews.development_crew import DevelopmentCrew
-
-# Import ceremony tracking - handle missing functions gracefully
-try:
-    from scripts.sprint_ceremony_tracker import SprintCeremonyTracker
-except ImportError:
-    SprintCeremonyTracker = None
-
-try:
-    from scripts.sprint_validator import SprintValidator
-except ImportError:
-    SprintValidator = None
-
-
-class SprintOrchestratorCrew:
-    """Sprint-level orchestration crew managing multiple development crews."""
-
-    def __init__(self):
-        """Initialize sprint orchestrator with management agents."""
-        self.registry = AgentRegistry()
-        self.agents = self._create_agents()
-        self.crew = self._create_crew()
-
-    def _create_agents(self) -> dict[str, Agent]:
-        """Create management agents for sprint orchestration."""
-        scrum_master_config = self.registry.get_agent("scrum-master")
-        devops_config = self.registry.get_agent("devops")
-
-        agents = {
-            "scrum_master": Agent(
-                role=scrum_master_config["role"],
-                goal="Orchestrate sprint execution with quality gates and process discipline",
-                backstory=scrum_master_config["backstory"],
-                verbose=True,
-                allow_delegation=True,  # Can delegate to development crews
-                llm=scrum_master_config["llm_client"],
-                tools=scrum_master_config["tools"],
-            ),
-            "devops": Agent(
-                role=devops_config["role"],
-                goal="Manage deployment pipeline and infrastructure during sprint",
-                backstory=devops_config["backstory"],
-                verbose=True,
-                allow_delegation=False,
-                llm=devops_config["llm_client"],
-                tools=devops_config["tools"],
-            ),
-        }
-
-        return agents
-
-    def _create_crew(self) -> Crew:
-        """Create the sprint orchestrator crew."""
-        return Crew(
-            agents=list(self.agents.values()),
-            tasks=[],  # Tasks created per sprint
-            process=Process.sequential,
-            verbose=True,
-            memory=False,  # Disable memory to avoid embeddings requirement
-        )
-
-    def _parse_sprint_md(self, sprint_number: int) -> dict[str, Any]:
-        """Parse SPRINT.md file and extract story information."""
-        sprint_file = Path("SPRINT.md")
-
-        if not sprint_file.exists():
-            raise FileNotFoundError("SPRINT.md not found. Sprint planning required.")
-
-        with open(sprint_file) as f:
-            content = f.read()
-
-        # Extract stories from SPRINT.md
-        # This is a simplified parser - in practice, you'd want more robust parsing
-        stories = []
-        lines = content.split("\n")
-
-        current_story = None
-        in_story_section = False
-
-        for line in lines:
-            line = line.strip()
-
-            # Look for story headers (e.g., "## Story 1: Add Authentication")
-            if line.startswith("## Story") and ":" in line:
-                if current_story:
-                    stories.append(current_story)
-
-                story_title = line.split(":", 1)[1].strip()
-                story_number = line.split(":")[0].replace("## Story", "").strip()
-
-                current_story = {
-                    "story_number": story_number,
-                    "story_title": story_title,
-                    "acceptance_criteria": [],
-                    "story_points": None,
-                    "priority": None,
-                    "status": "pending",
-                }
-                in_story_section = True
-
-            # Look for acceptance criteria
-            elif in_story_section and line.startswith("- ["):
-                ac = line.replace("- [ ]", "").replace("- [x]", "").strip()
-                if ac and current_story:
-                    current_story["acceptance_criteria"].append(ac)
-
-            # Look for story points
-            elif in_story_section and "story points:" in line.lower():
-                points = line.split(":")[-1].strip()
-                if current_story and points.isdigit():
-                    current_story["story_points"] = int(points)
-
-            # Look for priority
-            elif in_story_section and "priority:" in line.lower():
-                priority = line.split(":")[-1].strip()
-                if current_story:
-                    current_story["priority"] = priority
-
-            # End of story section
-            elif line.startswith("##") and not line.startswith("## Story"):
-                in_story_section = False
-
-        # Add the last story
-        if current_story:
-            stories.append(current_story)
-
-        return {
-            "sprint_number": sprint_number,
-            "stories": stories,
-            "total_points": sum(s.get("story_points", 0) for s in stories),
-            "total_stories": len(stories),
-        }
-
-    def _validate_sprint_readiness(self, sprint_number: int) -> dict[str, Any]:
-        """Validate that sprint is ready to start."""
-        dor_status = True  # Default to True if no ceremony tracker
-        structure_valid = True  # Default to True if no validator
-
-        # Check Definition of Ready if ceremony tracker is available
-        if SprintCeremonyTracker:
-            try:
-                ceremony_tracker = SprintCeremonyTracker()
-                dor_status = ceremony_tracker.can_start_sprint(sprint_number)
-            except Exception:
-                dor_status = True  # Fail gracefully
-
-        # Check sprint structure if validator is available
-        if SprintValidator:
-            try:
-                validator = SprintValidator()
-                structure_valid = validator.validate_sprint()
-            except Exception:
-                structure_valid = True  # Fail gracefully
-
-        return {
-            "definition_of_ready": dor_status,
-            "structure_valid": structure_valid,
-            "ready_to_start": dor_status and structure_valid,
-        }
-
-    def _validate_previous_sprint_closure(
-        self, previous_sprint_number: int
-    ) -> dict[str, Any]:
-        """Validate that the previous sprint is properly closed before starting new sprint."""
-        if previous_sprint_number <= 0:
-            # No previous sprint to validate (Sprint 1 case)
-            return {
-                "previous_sprint_closed": True,
-                "validation_reason": "No previous sprint to validate (Sprint 1 or earlier)",
-            }
-
-        try:
-            # Create sprint closure validation task for DevOps agent
-            sprint_closure_validation_task = Task(
-                description=f"""
-                **PREVIOUS SPRINT CLOSURE VALIDATION**: Validate Sprint {previous_sprint_number} is properly closed.
-
-                **Previous Sprint**: {previous_sprint_number}
-
-                **Your Validation Tasks**:
-                1. **Sprint Stories Completion Check**:
-                   - Verify all Sprint {previous_sprint_number} stories are marked complete
-                   - Check for any incomplete or blocked stories
-                   - Validate story completion meets Definition of Done
-                   - Report any stories that need carryover to next sprint
-
-                2. **Sprint Retrospective Validation**:
-                   - Verify Sprint {previous_sprint_number} retrospective was conducted
-                   - Check retrospective documentation exists
-                   - Validate lessons learned were captured
-                   - Ensure action items were identified for process improvement
-
-                3. **Sprint Metrics Finalization Check**:
-                   - Verify velocity was calculated for Sprint {previous_sprint_number}
-                   - Check burndown chart was generated
-                   - Validate completion rate and success metrics calculated
-                   - Ensure metrics were added to historical tracking
-
-                4. **GitHub Infrastructure Closure**:
-                   - Check Sprint {previous_sprint_number} milestone is closed
-                   - Verify GitHub project board is archived
-                   - Validate all issues are properly closed or moved
-                   - Check for any stale issues needing cleanup
-
-                5. **CI/CD Health Summary**:
-                   - Verify no critical build failures from Sprint {previous_sprint_number}
-                   - Check security scan results were addressed
-                   - Validate technical debt was documented
-                   - Ensure infrastructure is ready for next sprint
-
-                **Validation Rules**:
-                - All sprint stories must be complete or explicitly carried over
-                - Retrospective documentation is mandatory
-                - Sprint metrics must be finalized and archived
-                - GitHub infrastructure must be properly closed
-                - CI/CD pipeline must be healthy with no blocking issues
-
-                **Output Required**:
-                - Sprint closure status (CLOSED/INCOMPLETE)
-                - List of incomplete items preventing closure
-                - Required actions to complete closure
-                - Validation summary with specific findings
-                """,
-                agent=self.agents["devops"],
-                expected_output="Previous sprint closure validation complete with status, any incomplete items identified, and required actions for proper closure.",
-            )
-
-            # Execute sprint closure validation
-            validation_crew = Crew(
-                agents=[self.agents["devops"]],
-                tasks=[sprint_closure_validation_task],
-                process=Process.sequential,
-                verbose=True,
-                memory=False,
-            )
-
-            validation_result = validation_crew.kickoff()
-
-            # For now, we'll simulate validation results based on typical sprint closure checks
-            # In a real implementation, this would parse the actual validation results
-            validation_issues = []
-
-            # Check if SPRINT.md shows previous sprint as complete
-            try:
-                sprint_file = Path("SPRINT.md")
-                if sprint_file.exists():
-                    with open(sprint_file) as f:
-                        content = f.read()
-
-                    # Look for previous sprint section and completion markers
-                    if f"Sprint {previous_sprint_number}" not in content:
-                        validation_issues.append(
-                            f"Sprint {previous_sprint_number} not found in SPRINT.md"
-                        )
-
-                    # Check for retrospective section
-                    if f"Sprint {previous_sprint_number} Retrospective" not in content:
-                        validation_issues.append(
-                            f"Sprint {previous_sprint_number} retrospective not documented"
-                        )
-
-            except Exception as e:
-                validation_issues.append(f"Unable to validate SPRINT.md: {str(e)}")
-
-            # Determine if previous sprint is properly closed
-            is_closed = len(validation_issues) == 0
-
-            return {
-                "previous_sprint_closed": is_closed,
-                "previous_sprint_number": previous_sprint_number,
-                "validation_result": validation_result,
-                "validation_issues": validation_issues,
-                "required_actions": [
-                    f"Complete Sprint {previous_sprint_number} retrospective documentation",
-                    f"Finalize Sprint {previous_sprint_number} metrics and velocity",
-                    f"Close GitHub milestone for Sprint {previous_sprint_number}",
-                    f"Archive Sprint {previous_sprint_number} project board",
-                    "Address any incomplete stories or technical debt",
-                ]
-                if validation_issues
-                else [],
-                "closure_summary": {
-                    "stories_complete": len(validation_issues) == 0,
-                    "retrospective_done": "retrospective"
-                    not in " ".join(validation_issues).lower(),
-                    "metrics_finalized": True,  # Would be checked in real implementation
-                    "github_closed": True,  # Would be checked in real implementation
-                    "ci_cd_healthy": True,  # Would be checked in real implementation
-                },
-            }
-
-        except Exception as e:
-            return {
-                "previous_sprint_closed": False,
-                "previous_sprint_number": previous_sprint_number,
-                "error": str(e),
-                "required_actions": [
-                    f"Investigate Sprint {previous_sprint_number} closure status",
-                    "Manually verify previous sprint completion",
-                    "Complete any pending sprint closure tasks",
-                ],
-            }
-
-    def _setup_sprint_infrastructure(self, sprint_number: int) -> dict[str, Any]:
-        """Setup sprint infrastructure using DevOps agent."""
-        try:
-            # Create sprint setup task for DevOps agent
-            sprint_setup_task = Task(
-                description=f"""
-                **SPRINT INFRASTRUCTURE SETUP**: Prepare all infrastructure for Sprint {sprint_number} execution.
-
-                **Sprint Number**: {sprint_number}
-
-                **Your Tasks**:
-                1. **GitHub Project Board Setup**:
-                   - Create GitHub Projects v2 board for Sprint {sprint_number}
-                   - Configure custom fields (story points, priority, status)
-                   - Setup Kanban and Table views for sprint tracking
-                   - Initialize burndown chart tracking
-
-                2. **CI/CD Pipeline Health Check**:
-                   - Verify all GitHub Actions workflows are functioning
-                   - Check for any existing red builds that need fixing
-                   - Validate security scans are current and passing
-                   - Ensure test pipeline is ready for new commits
-
-                3. **Sprint Milestone Creation**:
-                   - Create GitHub milestone for Sprint {sprint_number}
-                   - Link all sprint stories to the milestone
-                   - Configure milestone due date and description
-
-                4. **Infrastructure Validation**:
-                   - Check Python environment and dependencies
-                   - Validate linting and type checking tools (ruff, mypy)
-                   - Ensure test coverage tracking is operational
-                   - Verify build cache and optimization settings
-
-                5. **Monitoring Setup**:
-                   - Initialize sprint metrics collection
-                   - Setup automated burndown snapshot capture
-                   - Prepare velocity tracking for completion
-                   - Configure CI/CD health monitoring
-
-                **Critical Rules**:
-                - All infrastructure must be green before sprint starts
-                - GitHub project board must be fully functional
-                - CI/CD pipeline must be healthy and ready
-                - Block sprint start if critical infrastructure issues found
-
-                **Output Required**:
-                - Sprint infrastructure status (READY/BLOCKED)
-                - GitHub project board URL and configuration
-                - CI/CD pipeline health summary
-                - List of any infrastructure issues found
-                """,
-                agent=self.agents["devops"],
-                expected_output="Sprint infrastructure setup complete with status, GitHub project board configured, CI/CD pipeline validated, and monitoring initialized.",
-            )
-
-            # Execute sprint setup
-            setup_crew = Crew(
-                agents=[self.agents["devops"]],
-                tasks=[sprint_setup_task],
-                process=Process.sequential,
-                verbose=True,
-                memory=False,
-            )
-
-            setup_result = setup_crew.kickoff()
-
-            return {
-                "status": "success",
-                "infrastructure_ready": True,
-                "github_project_created": True,
-                "ci_cd_healthy": True,
-                "monitoring_enabled": True,
-                "setup_result": setup_result,
-            }
-
-        except Exception as e:
-            return {"status": "error", "infrastructure_ready": False, "error": str(e)}
-
-    def _finalize_sprint_metrics(
-        self, sprint_number: int, sprint_data: dict, story_results: list[dict]
-    ) -> dict[str, Any]:
-        """Finalize sprint metrics and generate reports using DevOps agent."""
-        try:
-            # Create sprint finalization task for DevOps agent
-            sprint_finalization_task = Task(
-                description=f"""
-                **SPRINT FINALIZATION**: Generate final metrics and close Sprint {sprint_number} infrastructure.
-
-                **Sprint Number**: {sprint_number}
-                **Total Stories**: {sprint_data["total_stories"]}
-                **Total Points**: {sprint_data["total_points"]}
-
-                **Your Tasks**:
-                1. **Generate Sprint Metrics**:
-                   - Calculate final velocity (completed story points)
-                   - Generate completion rate statistics
-                   - Create burndown chart from daily snapshots
-                   - Compare actual vs planned velocity
-
-                2. **Update Velocity Tracking**:
-                   - Add Sprint {sprint_number} velocity to historical data
-                   - Update velocity trend charts (last 5 sprints)
-                   - Calculate predictability metrics (variance analysis)
-                   - Generate velocity improvement recommendations
-
-                3. **CI/CD Health Summary**:
-                   - Generate CI/CD uptime report for sprint period
-                   - Calculate build success rate and average build time
-                   - Report security scan results and coverage trends
-                   - Document any infrastructure improvements made
-
-                4. **GitHub Project Closure**:
-                   - Close GitHub milestone for Sprint {sprint_number}
-                   - Archive project board with final status
-                   - Generate project board health report
-                   - Update issue hygiene metrics
-
-                5. **Documentation and Handoff**:
-                   - Generate sprint retrospective data
-                   - Document lessons learned for infrastructure
-                   - Prepare velocity data for next sprint planning
-                   - Archive all sprint artifacts
-
-                **Output Required**:
-                - Final sprint velocity and completion metrics
-                - Burndown chart and velocity trend analysis
-                - CI/CD health summary for sprint period
-                - Archived project artifacts and documentation
-                """,
-                agent=self.agents["devops"],
-                expected_output="Sprint finalization complete with final metrics, velocity tracking updated, CI/CD health summarized, and all sprint artifacts archived.",
-            )
-
-            # Execute sprint finalization
-            finalization_crew = Crew(
-                agents=[self.agents["devops"]],
-                tasks=[sprint_finalization_task],
-                process=Process.sequential,
-                verbose=True,
-                memory=False,
-            )
-
-            finalization_result = finalization_crew.kickoff()
-
-            return {
-                "status": "success",
-                "metrics_finalized": True,
-                "velocity_updated": True,
-                "project_archived": True,
-                "finalization_result": finalization_result,
-            }
-
-        except Exception as e:
-            return {"status": "error", "metrics_finalized": False, "error": str(e)}
-
-    def _route_story_to_crew(self, story: dict[str, Any]) -> str:
-        """Determine which crew type should handle this story."""
-        # Simple routing logic - can be enhanced with ML/LLM classification
-        story_title = story.get("story_title", "").lower()
-
-        if any(keyword in story_title for keyword in ["test", "testing", "coverage"]):
-            return "testing_crew"
-        elif any(
-            keyword in story_title for keyword in ["deploy", "ci", "infrastructure"]
-        ):
-            return "devops_crew"
-        elif any(
-            keyword in story_title for keyword in ["refactor", "quality", "standards"]
-        ):
-            return "quality_crew"
-        else:
-            return "development_crew"  # Default for feature implementation
-
-    def _execute_story_crew(
-        self, story: dict[str, Any], crew_type: str
-    ) -> dict[str, Any]:
-        """Execute a single story using the appropriate crew."""
-        try:
-            if crew_type == "development_crew":
-                crew = DevelopmentCrew()
-                result = crew.kickoff(story)
-
-                # If story succeeded, validate CI/CD pipeline
-                if result["status"] == "success":
-                    ci_cd_result = self._validate_ci_cd_pipeline(story, result)
-                    result["ci_cd_validation"] = ci_cd_result
-
-                return {
-                    "story_number": story["story_number"],
-                    "story_title": story["story_title"],
-                    "crew_type": crew_type,
-                    "status": result["status"],
-                    "result": result,
-                }
-            else:
-                # Placeholder for other crew types
-                return {
-                    "story_number": story["story_number"],
-                    "story_title": story["story_title"],
-                    "crew_type": crew_type,
-                    "status": "not_implemented",
-                    "result": f"Crew type {crew_type} not yet implemented",
-                }
-
-        except Exception as e:
-            return {
-                "story_number": story["story_number"],
-                "story_title": story["story_title"],
-                "crew_type": crew_type,
-                "status": "error",
-                "error": str(e),
-            }
-
-    def _validate_ci_cd_pipeline(
-        self, story: dict[str, Any], dev_result: dict[str, Any]
-    ) -> dict[str, Any]:
-        """Validate CI/CD pipeline after story completion using DevOps agent."""
-        try:
-            # Create CI/CD validation task for DevOps agent
-            ci_cd_task = Task(
-                description=f"""
-                **CI/CD PIPELINE VALIDATION**: Validate build health and infrastructure after story completion.
-
-                **Story**: {story.get("story_title", "No title provided")}
-                **Story Number**: {story.get("story_number", "N")}
-
-                **Context**: Story has been implemented, reviewed, and committed to git. Validate that CI/CD pipeline is healthy.
-
-                **Your Tasks**:
-                1. **Build Validation**:
-                   - Check GitHub Actions workflow status
-                   - Verify all builds are passing (green status)
-                   - If red builds found, diagnose and report issues
-
-                2. **Test Pipeline Validation**:
-                   - Verify test suite execution completed successfully
-                   - Check test coverage maintained (target: >80%)
-                   - Validate no regression in test pass rate
-
-                3. **Security Scan Validation**:
-                   - Run Bandit security scan on changed files
-                   - Verify dependency security audit passed
-                   - Report any critical security vulnerabilities
-
-                4. **Quality Gate Validation**:
-                   - Verify linting (ruff) passed on all changed files
-                   - Verify type checking (mypy) passed
-                   - Check code formatting standards maintained
-
-                5. **Infrastructure Health Check**:
-                   - Validate build time is within acceptable range (<5 min)
-                   - Check resource utilization during build
-                   - Update build health metrics
-
-                **Critical Rules**:
-                - Red builds are P0 - must be fixed immediately
-                - Security vulnerabilities (critical) block story completion
-                - Test regression failures require immediate investigation
-                - Report specific file paths and error details
-
-                **Output Required**:
-                - CI/CD pipeline health status (PASS/FAIL/WARNING)
-                - Specific build/test/security results
-                - List of any issues found with resolution steps
-                - Updated pipeline health metrics
-                """,
-                agent=self.agents["devops"],
-                expected_output="CI/CD pipeline validation complete with health status, specific results for builds/tests/security, and any issues identified with resolution steps.",
-            )
-
-            # Execute CI/CD validation
-            validation_crew = Crew(
-                agents=[self.agents["devops"]],
-                tasks=[ci_cd_task],
-                process=Process.sequential,
-                verbose=True,
-                memory=False,
-            )
-
-            ci_cd_result = validation_crew.kickoff()
-
-            return {
-                "status": "success",
-                "pipeline_health": "PASS",  # This would be determined by actual validation
-                "validation_result": ci_cd_result,
-                "builds_passing": True,
-                "tests_passing": True,
-                "security_clear": True,
-                "quality_gates_passed": True,
-            }
-
-        except Exception as e:
-            return {
-                "status": "error",
-                "pipeline_health": "FAIL",
-                "error": str(e),
-                "builds_passing": False,
-                "tests_passing": False,
-                "security_clear": False,
-                "quality_gates_passed": False,
-            }
-
-    def _execute_stories_parallel(
-        self, stories: list[dict[str, Any]], max_workers: int = 3
-    ) -> list[dict[str, Any]]:
-        """Execute multiple stories in parallel using ThreadPoolExecutor."""
-        results = []
-
-        with ThreadPoolExecutor(max_workers=max_workers) as executor:
-            # Submit all stories for execution
-            future_to_story = {}
-
-            for story in stories:
-                crew_type = self._route_story_to_crew(story)
-                future = executor.submit(self._execute_story_crew, story, crew_type)
-                future_to_story[future] = story
-
-            # Collect results as they complete
-            for future in as_completed(future_to_story):
-                story = future_to_story[future]
-                try:
-                    result = future.result()
-                    results.append(result)
-                    print(f"✅ Completed: {story['story_title']}")
-                except Exception as e:
-                    error_result = {
-                        "story_number": story["story_number"],
-                        "story_title": story["story_title"],
-                        "status": "error",
-                        "error": str(e),
-                    }
-                    results.append(error_result)
-                    print(f"❌ Failed: {story['story_title']} - {e}")
-
-        return results
-
-    def _generate_sprint_metrics(
-        self, sprint_data: dict, story_results: list[dict]
-    ) -> dict[str, Any]:
-        """Generate sprint completion metrics."""
-        total_stories = len(story_results)
-        completed_stories = len([r for r in story_results if r["status"] == "success"])
-        failed_stories = len([r for r in story_results if r["status"] == "error"])
-
-        total_points_planned = sprint_data["total_points"]
-        completed_points = sum(
-            sprint_data["stories"][i].get("story_points", 0)
-            for i, result in enumerate(story_results)
-            if result["status"] == "success"
-        )
-
-        return {
-            "sprint_number": sprint_data["sprint_number"],
-            "completion_rate": (completed_stories / total_stories * 100)
-            if total_stories > 0
-            else 0,
-            "velocity": completed_points,
-            "stories_completed": completed_stories,
-            "stories_failed": failed_stories,
-            "total_stories": total_stories,
-            "points_completed": completed_points,
-            "points_planned": total_points_planned,
-            "success_rate": (completed_points / total_points_planned * 100)
-            if total_points_planned > 0
-            else 0,
-        }
-
-    def execute_sprint(
-        self, sprint_number: int, max_parallel_stories: int = 3
-    ) -> dict[str, Any]:
-        """
-        Execute an entire sprint using development crews.
-
-        Args:
-            sprint_number: Sprint number to execute
-            max_parallel_stories: Maximum number of stories to execute in parallel
-
-        Returns:
-            Sprint execution results with metrics
-        """
-        print(f"\n{'=' * 70}")
-        print(f"EXECUTING SPRINT {sprint_number} WITH CREWAI DEVELOPMENT TEAMS")
-        print(f"{'=' * 70}")
-
-        try:
-            # 0. Validate Previous Sprint Closure
-            print("🔍 Validating previous sprint closure...")
-            previous_sprint_validation = self._validate_previous_sprint_closure(
-                sprint_number - 1
-            )
-
-            if not previous_sprint_validation["previous_sprint_closed"]:
-                return {
-                    "status": "blocked",
-                    "sprint_number": sprint_number,
-                    "error": "Previous sprint not properly closed",
-                    "previous_sprint_validation": previous_sprint_validation,
-                    "required_actions": previous_sprint_validation.get(
-                        "required_actions", []
-                    ),
-                }
-
-            print("✅ Previous sprint closure validated")
-
-            # 1. DevOps Sprint Setup
-            print("🏗️ Setting up sprint infrastructure...")
-            sprint_setup = self._setup_sprint_infrastructure(sprint_number)
-
-            if sprint_setup["status"] != "success":
-                return {
-                    "status": "blocked",
-                    "sprint_number": sprint_number,
-                    "error": "Sprint infrastructure setup failed",
-                    "setup_result": sprint_setup,
-                }
-
-            print("✅ Sprint infrastructure setup complete")
-
-            # 2. Validate sprint readiness
-            print("📋 Validating sprint readiness...")
-            readiness = self._validate_sprint_readiness(sprint_number)
-
-            if not readiness["ready_to_start"]:
-                return {
-                    "status": "blocked",
-                    "sprint_number": sprint_number,
-                    "error": "Sprint not ready to start",
-                    "readiness_check": readiness,
-                }
-
-            print("✅ Sprint readiness validated")
-
-            # 2. Parse sprint definition
-            print("📖 Parsing SPRINT.md...")
-            sprint_data = self._parse_sprint_md(sprint_number)
-            print(
-                f"📊 Found {sprint_data['total_stories']} stories ({sprint_data['total_points']} points)"
-            )
-
-            # 3. Execute stories in parallel
-            print(f"🚀 Executing {len(sprint_data['stories'])} stories in parallel...")
-            story_results = self._execute_stories_parallel(
-                sprint_data["stories"], max_workers=max_parallel_stories
-            )
-
-            # 4. Generate metrics
-            print("📈 Generating sprint metrics...")
-            metrics = self._generate_sprint_metrics(sprint_data, story_results)
-
-            # 5. DevOps Sprint Finalization
-            print("🏁 Finalizing sprint infrastructure and metrics...")
-            finalization = self._finalize_sprint_metrics(
-                sprint_number, sprint_data, story_results
-            )
-
-            # 6. Return comprehensive results
-            return {
-                "status": "completed",
-                "sprint_number": sprint_number,
-                "sprint_data": sprint_data,
-                "story_results": story_results,
-                "metrics": metrics,
-                "sprint_setup": sprint_setup,
-                "sprint_finalization": finalization,
-                "execution_summary": {
-                    "total_stories": metrics["total_stories"],
-                    "completed_stories": metrics["stories_completed"],
-                    "completion_rate": f"{metrics['completion_rate']:.1f}%",
-                    "velocity": metrics["velocity"],
-                    "success_rate": f"{metrics['success_rate']:.1f}%",
-                    "infrastructure_ready": sprint_setup.get(
-                        "infrastructure_ready", False
-                    ),
-                    "ci_cd_healthy": all(
-                        sr.get("result", {})
-                        .get("ci_cd_validation", {})
-                        .get("pipeline_health")
-                        == "PASS"
-                        for sr in story_results
-                        if sr.get("status") == "success"
-                    ),
-                    "metrics_finalized": finalization.get("metrics_finalized", False),
-                },
-            }
-
-        except Exception as e:
-            return {
-                "status": "error",
-                "sprint_number": sprint_number,
-                "error": str(e),
-                "phase": "Sprint Orchestration",
-            }
-
-
-# Example usage for testing
-if __name__ == "__main__":
-    # Execute current sprint
-    orchestrator = SprintOrchestratorCrew()
-    result = orchestrator.execute_sprint(15)
-
-    print("\n" + "=" * 70)
-    print("SPRINT ORCHESTRATION COMPLETE")
-    print("=" * 70)
-
-    if result["status"] == "completed":
-        summary = result["execution_summary"]
-        print(f"Sprint {result['sprint_number']} Results:")
-        print(
-            f"  Stories: {summary['completed_stories']}/{summary['total_stories']} ({summary['completion_rate']})"
-        )
-        print(f"  Velocity: {summary['velocity']} story points")
-        print(f"  Success Rate: {summary['success_rate']}")
-
-        print("\nStory Results:")
-        for story_result in result["story_results"]:
-            status_emoji = "✅" if story_result["status"] == "success" else "❌"
-            print(
-                f"  {status_emoji} Story {story_result['story_number']}: {story_result['story_title']}"
-            )
-
-    else:
-        print(f"❌ Sprint execution failed: {result['error']}")
